+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [#2101](https://github.com/influxdata/kapacitor/issues/2101): Add multiple field support to the change detect node.
 - [#1961](https://github.com/influxdata/kapacitor/pull/1961): Add links to pagerduty2 alerts
 - [#1974](https://github.com/influxdata/kapacitor/issues/1974): Add additional metadata to Sensu alerts.
+- [#2154](https://github.com/influxdata/kapacitor/pull/2154): Add ability to skip ssl verification with an alert post node.
 
 ### Bugfixes